Use better file structure
