+++ tests/audit-tests/src/test/java/uk/gov/di/txma/audit/pages/IPVCoreStub/VisitCredentialIssuersPage.java
@@ -1,0 +1,18 @@
+package uk.gov.di.txma.audit.pages.IPVCoreStub;
+
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.FindBy;
+import org.openqa.selenium.support.PageFactory;
+import uk.gov.di.txma.audit.utilities.Driver;
+
+public class VisitCredentialIssuersPage {
+    public VisitCredentialIssuersPage() {
+        PageFactory.initElements(Driver.get(), this);
+    }
+    @FindBy(xpath = "//input[@value='KBV CRI Staging']")
+    public WebElement KBVCRIStaging;
+
+    @FindBy(xpath = "//input[@value='Address CRI Staging']")
+    public WebElement AddressCRIStaging;
+
+}
