+++ tests/audit-tests/src/test/java/uk/gov/di/txma/audit/pages/KBV/ExperianUATUserSearchResultsPage.java
@@ -1,24 +1,16 @@
-package pages;
-
-//import gov.di_ipv_kbv.utilities.Driver;
+package uk.gov.di.txma.audit.pages.KBV;
 
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
-import utilities.Driver;
+import uk.gov.di.txma.audit.utilities.Driver;
 
 public class ExperianUATUserSearchResultsPage {
     public ExperianUATUserSearchResultsPage() {
         PageFactory.initElements(Driver.get(), this);
     }
 
-    @FindBy(xpath = "//a[normalize-space()='Go to KBV CRI Build']")
-    public WebElement GoToKBVCRIBuild;
-
     @FindBy(xpath = "//a[normalize-space()='Go to KBV CRI Staging']")
     public WebElement GoToKBVCRIStaging;
-
-    @FindBy(xpath = "//a[normalize-space()='View KBV Answers']")
-    public WebElement ViewKBVAnswers;
 
 }
