+++ tests/audit-tests/src/test/java/uk/gov/di/txma/audit/pages/KBV/FindYourAddressPage.java
@@ -1,9 +1,9 @@
-package pages.KBV;
+package uk.gov.di.txma.audit.pages.KBV;
 
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
-import utilities.Driver;
+import uk.gov.di.txma.audit.utilities.Driver;
 
 public class FindYourAddressPage {
     public FindYourAddressPage() {
