+++ tests/audit-tests/src/test/java/uk/gov/di/txma/audit/pages/KBV/UserForKBVCRIStagingPage.java
@@ -1,11 +1,11 @@
-package pages;
+package uk.gov.di.txma.audit.pages.KBV;
 
-//import gov.di_ipv_kbv.utilities.Driver;
+//import gov.di_ipv_kbv.uk.gov.di.txma.audit.utilities.Driver;
 
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
-import utilities.Driver;
+import uk.gov.di.txma.audit.utilities.Driver;
 
 public class UserForKBVCRIStagingPage {
     public UserForKBVCRIStagingPage (){
