+++ tests/audit-tests/src/test/java/uk/gov/di/txma/audit/utilities/Driver.java
@@ -1,4 +1,4 @@
-package utilities;
+package uk.gov.di.txma.audit.utilities;
 
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.WebDriver;
